requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
google-generativeai==0.5.4
google-cloud-translate==3.15.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.0
pytest==8.1.1
pytest-asyncio==0.23.6
firebase-admin==6.5.0
slowapi==0.1.9
